requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
onnxruntime
python-multipart
